requirements.txt
agate==1.7.1
aiofiles==22.1.0
aiohttp==3.9.1
aiosignal==1.3.1
alembic==1.13.0
altair==4.2.0
anyio==3.7.1
appnope==0.1.3
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asn1crypto==1.5.1
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.1.0
azure-core==1.29.1
azure-identity==1.15.0
Babel==2.14.0
backcall==0.2.0
bcrypt==4.0.1
beautifulsoup4==4.12.2
betterproto==1.2.5
bleach==6.1.0
blinker==1.7.0
boto3==1.26.165
botocore==1.29.165
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.3
clickhouse-connect==0.6.23
clickhouse-driver==0.2.6
cloudpickle==3.0.0
colorama==0.4.6
comm==0.2.0
contourpy==1.2.0
croniter==1.3.7
cryptography==36.0.2
cycler==0.12.1
dask==2023.12.1
databricks-cli==0.18.0
datadog==0.44.0
dbt==1.0.0.35.1
dbt-core==1.7.4
dbt-extractor==0.5.1
dbt-semantic-interfaces==0.4.2
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
docker==6.1.3
duckdb==0.9.2
entrypoints==0.4
exceptiongroup==1.2.0
exchange_calendars==4.5.1
Faker==4.14.0
fastjsonschema==2.19.0
filelock==3.13.1
Flask==3.0.0
fonttools==4.47.0
fqdn==1.5.1
freezegun==1.2.2
frozenlist==1.4.1
fsspec==2023.12.2
future==0.18.3
gevent==23.9.1
gitdb==4.0.11
GitPython==3.1.34
google-api-core==2.15.0
google-auth==2.25.2
google-cloud-bigquery==3.14.1
google-cloud-core==2.4.1
google-cloud-dataproc==5.8.0
google-cloud-storage==2.14.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
great-expectations==0.15.50
greenlet==3.0.3
grpc-google-iam-v1==0.13.0
grpcio==1.60.0
grpcio-status==1.60.0
grpclib==0.4.6
gunicorn==21.2.0
h11==0.14.0
h2==4.1.0
hologram==0.0.16
hpack==4.0.0
httpcore==0.16.3
httpx==0.23.1
hyperframe==6.0.1
idna==3.6
importlib-metadata==6.11.0
inflection==0.5.1
ipykernel==6.15.0
ipython==7.34.0
ipython-genutils==0.2.0
ipywidgets==8.1.1
isodate==0.6.1
isoduration==20.11.0
itsdangerous==2.1.2
jaraco.classes==3.3.0
jedi==0.19.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.2
json5==0.9.14
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.9.0
jupyter-lsp==2.2.1
jupyter-server-proxy==3.2.1
jupyter_client==8.6.0
jupyter_core==5.5.1
jupyter_server==2.12.2
jupyter_server_terminals==0.5.1
jupyterlab==4.0.10
jupyterlab-widgets==3.0.9
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.2
kafka-python==2.0.2
keyring==24.3.0
kiwisolver==1.4.5
korean-lunar-calendar==0.3.1
lab==8.0
ldap3==2.9.1
leather==0.3.4
Levenshtein==0.23.0
locket==1.0.0
Logbook==1.5.3
lz4==4.3.2
mage-ai==0.9.55
makefun==1.15.2
Mako==1.3.0
Markdown==3.5.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
mashumaro==3.11
matplotlib==3.8.2
matplotlib-inline==0.1.6
mdurl==0.1.2
minimal-snowplow-tracker==0.0.2
mistune==3.0.2
mlflow==2.9.2
more-itertools==10.1.0
msal==1.26.0
msal-extensions==1.1.0
msgpack==1.0.7
multidict==6.0.4
nbclassic==1.0.0
nbclient==0.9.0
nbconvert==7.13.0
nbformat==5.9.2
nest-asyncio==1.5.8
networkx==2.8.8
newrelic==8.8.0
notebook==6.5.6
notebook_shim==0.2.3
numpy==1.26.2
oauthlib==3.2.2
overrides==7.4.0
packaging==23.2
pandas==1.5.3
pandas_market_calendars==4.3.3
pandocfilters==1.5.0
paramiko==3.4.0
parsedatetime==2.4
parso==0.8.3
partd==1.4.1
pathspec==0.10.3
pexpect==4.9.0
pickleshare==0.7.5
Pillow==10.1.0
platformdirs==3.11.0
polars==0.19.2
portalocker==2.8.2
postgres==4.0
prometheus-client==0.19.0
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==4.25.1
psutil==5.9.7
psycopg2==2.9.9
psycopg2-binary==2.9.3
psycopg2-pool==1.1
ptyprocess==0.7.0
pyarrow==10.0.1
pyarrow-hotfix==0.5
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==1.10.13
PyGithub==1.59.0
Pygments==2.17.2
PyJWT==2.6.0
pyluach==2.2.0
PyNaCl==1.5.0
pyodbc==4.0.39
pyOpenSSL==22.0.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-json-logger==2.0.7
python-Levenshtein==0.23.0
python-slugify==8.0.1
pytimeparse==1.1.8
pytz==2022.2.1
PyYAML==6.0.1
pyzmq==24.0.1
qtconsole==5.5.1
QtPy==2.4.1
querystring-parser==1.2.4
rapidfuzz==3.5.2
redis==5.0.1
referencing==0.32.0
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986==1.5.0
rfc3986-validator==0.1.1
rich==13.7.0
rpds-py==0.15.2
rsa==4.9
ruamel.yaml==0.17.17
s3transfer==0.6.2
scikit-learn==1.3.2
scipy==1.11.4
Send2Trash==1.8.2
sentry-sdk==1.19.1
shellingham==1.5.4
simpervisor==1.0.0
simplejson==3.19.2
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
snowflake-connector-python==3.6.0
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==1.4.50
sqlparams==6.0.1
sqlparse==0.4.3
sshtunnel==0.4.0
stringcase==1.2.0
tabulate==0.9.0
terminado==0.17.1
text-unidecode==1.3
thefuzz==0.19.0
threadpoolctl==3.2.0
tinycss2==1.2.1
tomli==2.0.1
tomlkit==0.12.3
toolz==0.12.0
tornado==6.4
tqdm==4.66.1
traitlets==5.14.0
trino==0.321.0
twelvedata==1.2.12
txt2tags==3.9
typer==0.9.0
types-python-dateutil==2.8.19.20240106
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==5.2
uri-template==1.3.0
urllib3==1.26.18
wcwidth==0.2.12
webcolors==1.13
webencodings==0.5.1
websocket==0.2.1
websocket-client==1.7.0
Werkzeug==3.0.1
widgetsnbextension==4.0.9
wrapt==1.16.0
yarl==1.9.4
zipp==3.17.0
zope.event==5.0
zope.interface==6.1
zstandard==0.22.0
